Move MinProtocolVersions to its own module
