+++ src/lib.rs
@@ -42,9 +42,6 @@ clippy::all
     bare_trait_objects
 )]
 
-pub(crate) mod pem;
-pub(crate) mod self_signed;
-
 /// Client configurations
 pub mod client;
 /// Server configurations
@@ -52,29 +49,11 @@ pub mod server;
 
 mod error;
 mod name;
+mod versions;
 
 pub use error::*;
 pub use name::*;
+pub use versions::*;
 
-/// Minimum protocol version allowed
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum MinProtocolVersion {
-    /// Allow TLS 1.2 and 1.3
-    V1_2,
-    /// Allow TLS 1.3 only
-    V1_3,
-}
-
-impl MinProtocolVersion {
-    pub(crate) fn versions(self) -> &'static [&'static rustls::SupportedProtocolVersion] {
-        static MIN_TLS12_VERSIONS: &[&rustls::SupportedProtocolVersion] =
-            &[&rustls::version::TLS13, &rustls::version::TLS12];
-        static MIN_TLS13_VERSIONS: &[&rustls::SupportedProtocolVersion] =
-            &[&rustls::version::TLS13];
-
-        match self {
-            Self::V1_2 => MIN_TLS12_VERSIONS,
-            Self::V1_3 => MIN_TLS13_VERSIONS,
-        }
-    }
-}
+pub(crate) mod pem;
+pub(crate) mod self_signed;
